requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
astroid==2.14.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
dill==0.3.6
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
platformdirs==3.0.0
pylint==2.16.2
python-dotenv==0.21.1
python-telegram-bot==20.1
rfc3986==1.5.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
yarl==1.8.2
